Added geometry helper functions to draw text

+++ SIM-VICUS/src/core3D/Vic3DGeometryHelpers.h
@@ -181,7 +181,7 @@ void addPolyLine(const std::vector<IBKMK::Vector3D> & polyline, const VICUS::Rot
 				 unsigned int & currentElementIndex, std::vector<VertexC> & vertexBufferData, std::vector<GLuint> & indexBufferData);
 
 
-void addText(const std::string & text, Qt::Alignment alignment, const unsigned int &textSize, const VICUS::RotationMatrix & matrix, const IBKMK::Vector3D &origin, const IBKMK::Vector2D &basePoint,
+void addText(const std::string & text, Qt::Alignment alignment, const double &textSize, const VICUS::RotationMatrix & matrix, const IBKMK::Vector3D &origin, const IBKMK::Vector2D &basePoint,
 			 double scalingFactor, double zScale, const QColor &color, unsigned int & currentVertexIndex, unsigned int & currentElementIndex, std::vector<Vertex> & vertexBufferData,
 			 std::vector<ColorRGBA> &colorBufferData, std::vector<GLuint> & indexBufferData);
